need fixs

+++ src/main/java/com/aiu/library/controller/WaitingListController.java
@@ -15,14 +15,14 @@ public class WaitingListController {
     private WaitingListService waitingListService;
 
     @PostMapping("/join/{bookId}/{memberId}")
-    public String join(@PathVariable Long bookId, @PathVariable int memberId, Model model) {
+    public String join(@PathVariable int bookId, @PathVariable int memberId, Model model) {
         String message = waitingListService.joinWaitingList(bookId, memberId);
         model.addAttribute("message", message);
         return "waiting-result";
     }
 
     @GetMapping("/view/{bookId}")
-    public String view(@PathVariable Long bookId, Model model) {
+    public String view(@PathVariable int bookId, Model model) {
         var queue = waitingListService.getWaitingList(bookId);
         model.addAttribute("bookId", bookId);
         model.addAttribute("waitingList", queue.getAll());
